modification classe voiture et client vers classe LOCATION

+++ garage/src/com/doignies/marie/Facture.java
@@ -9,18 +9,19 @@ public class Facture {
 	private Client client;
 	private Voiture voiture;
 	private LocalDate date;
-	private int prixLocation;
+	private long prixLocation;
 
 	// /////////////////////
 	// *** Constructeurs ***
 	// /////////////////////
-	public Facture(final Client newClient, final Voiture newVoiture, final int newPrixLocation) {
+	public Facture(final Client newClient, final Voiture newVoiture, final long newPrixLocation) {
 		// On ne met pas newDate
 		// dans la liste des parametres car lors de la Création d'une facture
 		// 1. On met la date du jour à l'instant T
 		this.client = newClient;
 		this.voiture = newVoiture;
 		this.date = LocalDate.now();
+		// TODO : ICI il faut la date de fin de la location
 		this.prixLocation = newPrixLocation;
 	}
